requirements.txt
google-generativeai>=0.3.0
google-genai
pillow>=10.0.0
PyMuPDF>=1.23.0
tenacity>=8.2.0
biopython>=1.81
python-dotenv>=1.0.0
pandoc>=2.3
tectonic>=0.12.0
flask>=2.0.0
werkzeug>=2.0.0
openai>=1.0.0
